requirements.txt
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
black==19.3b0
Click==7.0
decorator==4.4.0
Django==2.2.3
django-debug-toolbar==2.0
django-filter==2.2.0
django-jsonfield==1.2.0
djangorestframework==3.10.1
ipdb==0.12.1
ipython==7.7.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.14.1
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.3.1
pytz==2019.1
rope==0.14.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
traitlets==4.3.2
typed-ast==1.4.0
wcwidth==0.1.7
wrapt==1.11.2
